Extract orchestration logic from `processor.go`.
